requirements-mac.txt
# Requirements for Apple Silicon (Mac M1/M2/M3/M4)
# Use with: pip install -r requirements-mac.txt

# PyTorch with MPS support (install separately first)
# torch>=2.0.0
# torchvision>=0.15.0

# Core dependencies
transformers==4.46.3
tokenizers==0.20.3

# Web service
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# PDF processing
PyMuPDF>=1.23.0

# Image processing
Pillow>=10.0.0

# Model dependencies
einops>=0.7.0
addict>=2.4.0
easydict>=1.10

# Visualization
matplotlib>=3.8.0

# Numerical computing
numpy<2.0,>=1.23.2

# Optional: ModelScope fallback
modelscope>=1.9.0
